17.3, added remove item button, added icons, added r to lpgin page

+++ src/components/Sound/tools.js
@@ -1,5 +1,7 @@
-import { PermissionsAndroid } from 'react-native';
+import 'react-native-get-random-values';
+import { PermissionsAndroid, Platform } from 'react-native';
 import RNFS from 'react-native-fs'
+import { v4 as uuidv4 } from 'uuid';
 
 import { musicDirectory } from '../../constants'
 
@@ -13,10 +15,11 @@ const hasAndroidStoragePermission = async () => {
     const status = await PermissionsAndroid.requestMultiple(permission);
     return status === 'granted';
 }
-export const compileNewPathToSound = async (title, originUrl) => {
-    // console.log(originUrl)
-    if (await hasAndroidStoragePermission()) {
+export const createNewSound = async (title, originUrl) => {
+// ULTRA SUPER POWER MEGA GALAXY GAMING PLUS PREMIAL function
+    if (Platform.OS === 'android' && await hasAndroidStoragePermission()) {
         try {
+            const id = uuidv4()
             try {
                 const res = await RNFS.readDir(musicDirectory)
                 console.log('musicDirectory:', res)
@@ -27,18 +30,26 @@ export const compileNewPathToSound = async (title, originUrl) => {
                 const fileResult = await RNFS.readFile(originUrl, 'base64')
                 try {
                     let path = (musicDirectory + title.replaceAll(/[ -]/g, '_')).toLocaleLowerCase() // 
-                    await RNFS.writeFile(path, fileResult, 'base64')
-                    console.log('writen')
-                    return { title, url: path }
+                    const exist = await RNFS.exists(path)
+                    if (exist) {
+                        console.log('File already exist:', path)
+                        return ['File already added', null]
+                    } else {
+                        await RNFS.writeFile(path, fileResult, 'base64')
+                        console.log('writen')
+                        return [null, { title, url: path, id }]
+                    }
+
                 } catch (err) {
                     console.log('fail to write:', err)
                 }
             } catch (err) {
-                console.log('doesn`t readfile', err)
+                console.log('didn`t read file', err)
+                return ['Cannot read file', null]
             }
         } catch (err) {
             console.log('err:', err)
         }
     }
-    return false
+    return ['Error', null]
 }
